Add pwnlib.protocols module
Migrate pwnlib.adb.protocol into that directory, and add compatibility shims
